# Optimize large Top N queries